Improve repository type validation message but printing the valid options
