+++ scm-core/src/main/java/sonia/scm/validation/RepositoryType.java
