+++ scm-webapp/src/test/java/sonia/scm/validation/ValidationModuleTest.java
@@ -43,8 +43,8 @@ import javax.validation.Validator;
 import java.util.Collections;
 import java.util.Set;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.when;
+import static org.assertj.core.api.Assertions.*;
+import static org.mockito.Mockito.*;
 
 class ValidationModuleTest {
